trying functions in one file
